Update student_agent.py
I turned random move into a separate function and called it before every step, so we don't return an empty value. We can get rid of it but I don't think it takes up much space

I added an if statement in step if our minimax recursive call does not return any non-zero values for its max_evaluation. This allows us to pick a better-than-random move even if our minimax doesn't do anything. The function iterates over our possible moves and compares the # of moves we could make and what our opponent could make if we took that move. It is not entirely accurate because our moves our given before we check our opponents but it's relatively fast and wins us many more games. 

I also change the starting mineval and maxeval in minimax from -5 and 5 to 0. Since most moves return a score of 0 (unless we have a super big depth) it doesn't always update us to a right or left move.

As of now our step:

Calls a random move (we can get rid of)
Calls our minimax (doesn't do anything for most moves unless a) depth is super long or b) we are very close to winning)
If minimax doesn't do anything, we use the iterative nummoves function: this is providing most of our results rn